# src/main/java/com/amir/eventmanager/location/web/ErrorMessageResponse.java
package com.amir.eventmanager.location.web;

import java.time.LocalDateTime;

public record ErrorMessageResponse(
        String message,
        String detailedMessage,
        LocalDateTime dateTime
) {

}