# src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  description: API for managing event locations
  title: Event Management Service
  version: 1.0.0
servers:
- url: http://localhost:8080
security:
- bearerAuth: []
...
paths:
  /locations:
    get:
      description: Get all locations. Available to all authorized users
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/EventLocation'
                type: array
          description: List of all locations
        "401":
          content:
            application/json:
              example:
                message: Authentication required
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Authentication is missing or failed
        "403":
          content:
            application/json:
              example:
                message: Insufficient permissions to perform the operation
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: User does not have sufficient permissions to perform this operation
        "500":
          content:
            application/json:
              example:
                message: Internal server error
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Internal server error
      summary: "Get list of all locations. Allowed roles=[ADMIN, USER]"
      tags:
      - Locations
    post:
      description: Create a new location. Available only to admins. Id must be set on the application side (by database)
      requestBody:
        content:
          application/json:
            example:
              name: Ice Palace
              address: "St. Petersburg, Pyatiletok 1"
              capacity: 12300
              description: Sports and concert complex in St. Petersburg
            schema:
              $ref: '#/components/schemas/EventLocation'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventLocation'
          description: Location successfully created
        "400":
          content:
            application/json:
              example:
                message: Invalid request
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: "Request with invalid data. For example, attempt to use an already taken login or invalid request parameters."
        "401":
          content:
            application/json:
              example:
                message: Authentication required
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Authentication is missing or failed
        "403":
          content:
            application/json:
              example:
                message: Insufficient permissions to perform the operation
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: User does not have sufficient permissions to perform this operation
        "404":
          content:
            application/json:
              example:
                message: Entity not found
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Entity not found
        "500":
          content:
            application/json:
              example:
                message: Internal server error
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Internal server error
      summary: "Create new location. Allowed roles=[ADMIN]"
      tags:
      - Locations
  /locations/{locationId}:
    delete:
      description: "Delete location by ID. Cannot delete locations that already have scheduled events (should return 400 error)"
      parameters:
      - description: Unique location identifier
        in: path
        name: locationId
        required: true
        schema:
          example: 42
          format: int64
          type: integer
      responses:
        "204":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventLocation'
          description: Location successfully deleted
        "401":
          content:
            application/json:
              example:
                message: Authentication required
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Authentication is missing or failed
        "403":
          content:
            application/json:
              example:
                message: Insufficient permissions to perform the operation
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: User does not have sufficient permissions to perform this operation
        "404":
          content:
            application/json:
              example:
                message: Entity not found
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Entity not found
        "500":
          content:
            application/json:
              example:
                message: Internal server error
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Internal server error
      summary: "Delete location. Allowed roles=[ADMIN]"
      tags:
      - Locations
    get:
      description: "Get location by ID. Consider what response codes should be returned in different scenarios. For example, 404 code if entity not found."
      parameters:
      - description: Unique location identifier
        in: path
        name: locationId
        required: true
        schema:
          example: 42
          format: int64
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventLocation'
          description: Successful response
        "400":
          content:
            application/json:
              example:
                message: Invalid request
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: "Request with invalid data. For example, attempt to use an already taken login or invalid request parameters."
        "401":
          content:
            application/json:
              example:
                message: Authentication required
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Authentication is missing or failed
        "403":
          content:
            application/json:
              example:
                message: Insufficient permissions to perform the operation
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: User does not have sufficient permissions to perform this operation
        "404":
          content:
            application/json:
              example:
                message: Entity not found
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Entity not found
        "500":
          content:
            application/json:
              example:
                message: Internal server error
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Internal server error
      summary: "Get location by ID. Allowed roles=[ADMIN, USER]"
      tags:
      - Locations
    put:
      description: "Update location by its ID. Available only to admins. Consider parameter validation. Also consider that location capacity cannot be reduced because there may already be events scheduled at this location."
      parameters:
      - description: Unique location identifier
        in: path
        name: locationId
        required: true
        schema:
          example: 42
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            example:
              name: New location name
              address: New address
              capacity: 20000
              description: New description
            schema:
              $ref: '#/components/schemas/EventLocation'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventLocation'
          description: Location successfully updated
        "400":
          content:
            application/json:
              example:
                message: Invalid request
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: "Request with invalid data. For example, attempt to use an already taken login or invalid request parameters."
        "401":
          content:
            application/json:
              example:
                message: Authentication required
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Authentication is missing or failed
        "403":
          content:
            application/json:
              example:
                message: Insufficient permissions to perform the operation
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: User does not have sufficient permissions to perform this operation
        "404":
          content:
            application/json:
              example:
                message: Entity not found
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Entity not found
        "500":
          content:
            application/json:
              example:
                message: Internal server error
                detailedMessage: More detailed error message
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Internal server error
      summary: "Update location. Allowed roles=[ADMIN]"
      tags:
      - Locations
...
components:
...
  schemas:
    ErrorMessageResponse:
      example:
        message: Error message
        detailedMessage: More detailed error message
        dateTime: 2024-12-03T10:15:30
      properties:
        message:
          description: Human-readable error message
          type: string
        detailedMessage:
          description: "More detailed error message. For example, exception content"
          type: string
        dateTime:
          description: Date and time of error occurrence. Format "YYYY-MM-DDThh:mm:ss"
          format: date-time
          type: string
      required:
      - dateTime
      - detailedMessage
      - message
      type: object
    EventLocation:
      example:
        id: 42
        name: Ice Palace
        address: "St. Petersburg, Pyatiletok 1"
        capacity: 12300
        description: Sports and concert complex in St. Petersburg
      properties:
        id:
          description: Unique location identifier
          format: int64
          readOnly: true
          type: integer
        name:
          description: Location name
          type: string
        address:
          description: Location address
          type: string
        capacity:
          description: Location capacity
          minimum: 5
          type: integer
        description:
          description: Additional location description
          type: string
      required:
      - address
      - capacity
      - name
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http